* test/common/tests.c (main): Fix compiler warnings.

+++ neon/trunk/test/common/tests.c
@@ -242,8 +242,8 @@ int main(int argc, char *argv[])
             printf("%2d. %s%.*s ", n, test_name, 
                    (int) (strlen(dots) - strlen(test_name)), dots);
         else {
-            printf("\r%s%.*s %u/%lu ", test_suite, 
-                   strlen(dots) - strlen(test_suite), dots,
+            printf("\r%s%.*s %u/%u ", test_suite, 
+                   (int) (strlen(dots) - strlen(test_suite)), dots,
                    n + 1, count);
             fflush(stdout);
         }
@@ -360,8 +360,8 @@ int main(int argc, char *argv[])
 
 	reap_server();
             
-        printf("\r%s%.*s %u/%lu ", test_suite, 
-               strlen(dots) - strlen(test_suite), dots,
+        printf("\r%s%.*s %u/%u ", test_suite, 
+               (int) (strlen(dots) - strlen(test_suite)), dots,
                n + 1, count);
     }
 
@@ -380,8 +380,8 @@ int main(int argc, char *argv[])
             printf("<- all tests skipped for `%s'.\n", test_suite);
     } else {
         if (quiet) {
-            printf("\r%s%.*s %u/%lu ", test_suite, 
-                   strlen(dots) - strlen(test_suite), dots,
+            printf("\r%s%.*s %u/%u ", test_suite, 
+                   (int) (strlen(dots) - strlen(test_suite)), dots,
                    passes, count);
             if (fails == 0) {
                 COL("32");
